Connecting API to GUI, Homepage and single view page

+++ src/main/java/com/example/application/utils/SingleItem.java
@@ -2,6 +2,7 @@ package com.example.application.utils;
 
 import com.example.application.views.SingleResultView;
 import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.Html;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.html.Div;
@@ -14,21 +15,25 @@ public class SingleItem extends Div{
 	 * 
 	 */
 	private static final long serialVersionUID = -3268436718729722814L;
-
+	
 	public SingleItem(){
         setClassName("single-item");
     }
 
-    public Component init(String name, String image, boolean sponsored, Integer id){
+	
+	//Initialization of one destination in result view 
+    public Component init(String name, String image, boolean sponsored, Long id){
         H1 title = new H1(name);
         title.setClassName("title");
         getElement().getStyle().set("background-image", "url(images/"+image+")");
         addClickListener(e->{
+        	//navigation to single result page and passing ID of destination
             UI.getCurrent().navigate(SingleResultView.class, id);
         });
         if(sponsored){
             
         }
+
         add(title);
         return this;
     }
